requirements.txt
mcp>=1.11.0
requests>=2.31.0
langchain-core>=0.1.0 